--- components/RollControl.vue
@@ -1,13 +1,0 @@
-<template>
-    <div class="p-4 bg-slate-700 rounded-2xl">
-        <TextElement>Custom Roll</TextElement>
-    </div>
-</template>
-
-<script>
-import TextElement from './content/TextElement.vue'
-export default {
-    name: 'RollControl',
-    components: { TextElement },
-}
-</script>
