--- components/layout/Column.vue
@@ -1,7 +1,0 @@
-<template>
-    <div><slot /></div>
-</template>
-
-<script>
-export default { name: 'Column' }
-</script>
