--- components/layout/LayoutDivider.vue
@@ -1,9 +1,0 @@
-<template>
-    <div class="relative z-10">
-        <div class="absolute w-full h-px bg-divider" />
-    </div>
-</template>
-
-<script>
-export default {}
-</script>
