--- layouts/default.vue
@@ -1,19 +1,0 @@
-<template>
-    <Columns>
-        <Column>
-            <Nuxt />
-        </Column>
-        <Column></Column>
-    </Columns>
-</template>
-
-<script>
-import Column from '~/components/layout/Column.vue'
-import Columns from '~/components/layout/Columns.vue'
-export default {
-    components: {
-        Column,
-        Columns,
-    },
-}
-</script>
