+++ src/components/layout/Column.vue
@@ -1,0 +1,9 @@
+<template>
+	<div><slot /></div>
+</template>
+
+<script>
+import Vue from 'vue'
+
+export default Vue.extend({ name: 'Column' })
+</script>
