+++ src/components/layout/LayoutDivider.vue
@@ -1,0 +1,11 @@
+<template>
+	<div class="relative z-10">
+		<div class="absolute h-px w-full bg-divider" />
+	</div>
+</template>
+
+<script>
+import Vue from 'vue'
+
+export default Vue.extend({})
+</script>
