+++ src/pages/index.vue
@@ -1,12 +1,12 @@
 <template>
-    <div class="p-20">
-        <header>
-            <Heading :level="1">Dice Master</Heading>
-        </header>
-        <LayoutStack space="4" align="right">
-            <roll-control />
-        </LayoutStack>
-    </div>
+	<div class="p-20">
+		<header>
+			<Heading :level="1">Dice Master</Heading>
+		</header>
+		<LayoutStack space="4" align="right">
+			<roll-control />
+		</LayoutStack>
+	</div>
 </template>
 
 <script lang="ts">
@@ -14,12 +14,13 @@ import Vue from 'vue'
 import Heading from '~/components/content/Heading.vue'
 import LayoutStack from '~/components/layout/LayoutStack.vue'
 import RollControl from '~/components/RollControl.vue'
+
 export default Vue.extend({
-    components: {
-        Heading,
-        RollControl,
-        LayoutStack,
-    },
+	components: {
+		Heading,
+		RollControl,
+		LayoutStack,
+	},
 })
 </script>
